Fix , integrated linter/typescript and fixed files and static imgs types
