+++ src/main.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./App.css";
-import { App } from "./App.jsx";
+import { App } from "./App";
 import "./index.css";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { InvisibleFriend } from "./pages/invisibleApp/InvisibleFriend.jsx";
 import { NumberAsigner } from "./pages/numberAsigner/invisibleApp/NumberAsigner.jsx";
 import TicToe from "./pages/TicTacToe/TicToe.jsx";
@@ -17,7 +17,9 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
         <Route path="/NumberAsigner" element={<NumberAsigner />} />
         <Route
           path="/GabrielaPortfolio"
-          to={"https://gabrielacandio.netlify.app/"}
+          element={
+            <Navigate to="https://gabrielacandio.netlify.app/" replace />
+          }
         />
         <Route path="/TicTacToe" element={<TicToe />} />
       </Routes>
