+++ src/pages/TicTacToe/tictac1.tsx
